requirements.txt
pandas 
sqlalchemy 
psycopg2-binary
pgadmin4